We've made some mistakes
I get it. We had a month and a half to complete this project. We got started nice and early; the prototype and project proposal were completed well within the specified timeframe.

And yet.

After three weeks of waiting for four different amazon orders to arrive and a week or two at the end of October where a project due in a month didn't quite make the priority list among the myriad of math, physics, and cs projects due imminently, we found ourselves short on time. We had a week to put the project together; not an impossible task, certainly. Not even a difficult one. However, we got ourselves into a run-in with the law. Murphy's law. The gyroscope (hastily soldered in a tiny office on the third floor of the MC building, across from MATHSOC) gave us more trouble than it was worth, and the LCD display (soldered in the middle of an MC hallway as a MATHSOC meeting was let out) absolutely refused to work. We were able, after some finnicking, to get the backlight to light up, and after a full evening of searching Arduino forums we figured out which ports to connect the display to on the sensor shield, but getting any characters to show up was, for us, essentially impossible. Manually adjusting the contrast of the display using a makeshift screwdriver put together out of spare wires didn't work, nor did a veteran arduino user's "absolutely guaranteed method to connect to an I2C display." Eventually the display just broke entirely and not even the backlight would light up. Not too much of a hurdle, of course; except that this LCD was the most critical element of our project and there were now only hours until the due date.

But we're Waterloo students. The unofficial motto of Waterloo is to go "beyond ideas."

We decided to take this to heart and go beyond our original idea, completely scrapping our LCD-reliant Heads Up game in favor of a game controller that could be put together and linked to Unity with just the speakers, gyroscope, and sensor shield we had working. So if you're wondering why we're committing an entirely new project to the Git Repo less than ten hours before our demo, now you know.